+++ src/main/kotlin/com/example/address_project/domain/road_code/service/FileReaderRoadCodeEnglishService.kt
@@ -1,5 +1,6 @@
 package com.example.address_project.domain.road_code.service
 
+import com.example.address_project.domain.facade.SaveFileFacade
 import com.example.address_project.domain.road_code.domain.RoadCode
 import com.example.address_project.domain.road_code.domain.repository.RoadCodeRepository
 import com.example.address_project.infrastructure.common.feign.dto.UnzipFile
@@ -13,36 +14,20 @@ import java.io.File
 @Service
 class FileReaderRoadCodeEnglishService (
 
-    @Value("\${upload.dir}")
-    private var uploadDir: String,
 
     private val roadCodeRepository: RoadCodeRepository,
-
+    private val saveFileFacade: SaveFileFacade,
     private val addressZipFileService: AddressZipFileService
 ) {
 
     @Transactional
     fun readerEnglish(file: MultipartFile, unzipFile: UnzipFile) {
 
-        val filePath = saveFile(file)
+        val filePath = saveFileFacade.saveFile(file)
 
         addressZipFileService.addressFileWriter(unzipFile)
 
         saveEnglishRoadCode(filePath)
-    }
-
-    private fun saveFile(file: MultipartFile): String {
-        val uploadPath = File(uploadDir)
-
-        if (!uploadPath.exists()) {
-            uploadPath.mkdirs()
-        }
-
-        val targetLocation = File(uploadPath, file.originalFilename!!)
-
-        file.transferTo(targetLocation)
-
-        return targetLocation.absolutePath
     }
 
     private fun saveEnglishRoadCode(filePath: String) {
