+++ src/main/kotlin/com/example/address_project/domain/street_number/service/FileReaderStreetNumberEnglishService.kt
@@ -1,5 +1,6 @@
 package com.example.address_project.domain.street_number.service
 
+import com.example.address_project.domain.facade.SaveFileFacade
 import com.example.address_project.domain.street_number.domain.StreetNumber
 import com.example.address_project.domain.street_number.domain.repository.StreetNumberRepository
 import com.example.address_project.infrastructure.common.feign.dto.UnzipFile
@@ -13,36 +14,19 @@ import java.io.File
 @Service
 class FileReadeStreetNumberEnglishService (
 
-        @Value("\${upload.dir}")
-        private var uploadDir: String,
-
         private val streetNumberRepository : StreetNumberRepository,
-
+        private val saveFileFacade: SaveFileFacade,
         private val addressZipFileService: AddressZipFileService
 ) {
 
     @Transactional
     fun readerEnglish(file: MultipartFile, unzipFile: UnzipFile) {
 
-        val filePath = saveFile(file)
+        val filePath = saveFileFacade.saveFile(file)
 
         addressZipFileService.addressFileWriter(unzipFile)
 
         saveEnglishStreetNumber(filePath)
-    }
-
-    private fun saveFile(file: MultipartFile): String {
-        val uploadPath = File(uploadDir)
-
-        if (!uploadPath.exists()) {
-            uploadPath.mkdirs()
-        }
-
-        val targetLocation = File(uploadPath, file.originalFilename!!)
-
-        file.transferTo(targetLocation)
-
-        return targetLocation.absolutePath
     }
 
     private fun saveEnglishStreetNumber(filePath: String) {
