+++ src/main/kotlin/com/example/address_project/domain/street_number/service/FileReaderStreetNumberService.kt
@@ -1,5 +1,6 @@
 package com.example.address_project.domain.street_number.service
 
+import com.example.address_project.domain.facade.SaveFileFacade
 import com.example.address_project.domain.street_number.domain.StreetNumber
 import com.example.address_project.domain.street_number.domain.repository.StreetNumberRepository
 import com.example.address_project.infrastructure.common.feign.dto.UnzipFile
@@ -13,34 +14,19 @@ import java.io.File
 @Service
 class FileReaderStreetNumberService (
 
-        @Value("\${upload.dir}")
-        private var uploadDir: String,
         private val roadCodeRepository: StreetNumberRepository,
+        private val saveFileFacade: SaveFileFacade
         private val addressZipFileService: AddressZipFileService
 ) {
 
     @Transactional
     fun readerKorea(reqType: String, year: String, month: String, fileName: String, realFileName: String, file: MultipartFile, unzipFile: UnzipFile) {
 
-        val filePath = saveFile(file)
+        val filePath = saveFileFacade.saveFile(file)
 
         addressZipFileService.addressFileWriter(unzipFile)
 
         saveStreetNumber(filePath)
-    }
-
-    private fun saveFile(file: MultipartFile): String {
-        val uploadPath = File(uploadDir)
-
-        if (!uploadPath.exists()) {
-            uploadPath.mkdirs()
-        }
-
-        val targetLocation = File(uploadPath, file.originalFilename!!)
-
-        file.transferTo(targetLocation)
-
-        return targetLocation.absolutePath
     }
 
     private fun saveStreetNumber(filePath: String) {
